+++ sPHPf/vendors/coldstarstudios/databases/Pagination.php
@@ -72,12 +72,12 @@ class Pagination{
         
         if($this->start_num >= $this->reg_num){
             $fl = $this->start_num - $this->reg_num;
-            $pagination .= "<input type=\"button\" onclick=\"window.location='$this->page$separator$this->var_name=$fl';\" value=\"$this->before_term\">";
+            $pagination .= "<input type=\"button\" class=\"btn\" onclick=\"window.location='$this->page$separator$this->var_name=$fl';\" value=\"$this->before_term\">";
         }
         
-        if($this->start_num + $this->reg_num + $results_in_page < $this->total){
+        if($this->start_num + $this->reg_num + $results_in_page <= $this->total){
             $fr = $this->start_num + $this->reg_num;
-            $pagination .= "<input type=\"button\" onclick=\"window.location='$this->page$separator$this->var_name=$fr';\" value=\"$this->next_term\">";
+            $pagination .= "<input type=\"button\" class=\"btn\" onclick=\"window.location='$this->page$separator$this->var_name=$fr';\" value=\"$this->next_term\">";
         }
         return $pagination;
     }